Plug-in files moved
