+++ plugins/org.eclipse.gemoc.ale.language.metaprogramming/plugin.xml
@@ -5,7 +5,7 @@
          point="org.eclipse.gemoc.gemoc_language_workbench.metaprog">
       <approach
             name="ale"
-            validator="metaprogramming.aleapproach.AleRuleProvider">
+            validator="org.eclipse.gemoc.ale.language.metaprogramming.AleRuleProvider">
       </approach>
    </extension>
