+++ plugins/org.eclipse.gemoc.ale.language.metaprogramming/src/org/eclipse/gemoc/ale/language/metaprogramming/AleRule.java
@@ -1,4 +1,4 @@
-package rules;
+package org.eclipse.gemoc.ale.language.metaprogramming;
 
 import java.util.ArrayList;
 import java.util.List;
